Correct type warning
Resolves this error:

In file included from src/main.cpp:18:
lib/BQ27427/src/BQ27427.h:34:28: warning: unsigned conversion from 'int' to 'uint8_t' {aka 'unsigned char'} changes value from '12848' to '48' [-Woverflow]
   34 |           uint8_t chemId = BQ27427_CHEM_ID_A); // Lithium Cobalt Oxide (LiCoO₂)
      |                            ^~~~~~~~~~~~~~~~~
lib/BQ27427/src/BQ27427.h:34:28: warning: unsigned conversion from 'int' to 'uint8_t' {aka 'unsigned char'} changes value from '12848' to '48' [-Woverflow]
Compiling .pio/build/esp32-c3-devkitm-1/lib6b3/BQ27427/BQ27427.cpp.o
In file included from lib/BQ27427/src/BQ27427.cpp:14:
lib/BQ27427/src/BQ27427.h:34:28: warning: unsigned conversion from 'int' to 'uint8_t' {aka 'unsigned char'} changes value from '12848' to '48' [-Woverflow]
   34 |           uint8_t chemId = BQ27427_CHEM_ID_A); // Lithium Cobalt Oxide (LiCoO₂)

+++ src/BQ27427.h
@@ -31,7 +31,7 @@ public:
 					uint16_t designVoltage = 4200, 
 					uint16_t taperCurrent = 20, 
 					uint16_t terminateVoltage = 3000,
-          uint8_t chemId = BQ27427_CHEM_ID_A); // Lithium Cobalt Oxide (LiCoO₂)
+          uint16_t chemId = BQ27427_CHEM_ID_A); // Lithium Cobalt Oxide (LiCoO₂)
 	
 	/**
 	    Initializes I2C and verifies communication with the BQ274xx.
